backend/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
Django==4.2.1
django-extensions==3.2.1
django-filter==23.2
django-solo==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
idna==3.4
importlib-metadata==6.6.0
ipdb==0.13.13
ipython==8.12.2
jedi==0.18.2
Markdown==3.4.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing-extensions==4.6.2
urllib3==2.0.2
wcwidth==0.2.6
zipp==3.15.0
django-cors-headers==4.0.0
